Feature/storybook 02 (#59)
* docs: Add storybook for hooks

* refactor: Dialog Component props are duplicated with dialog tag`s attributes, so remove them.
dialog tag onClick → onClickAway

* docs: Change storybook Story name
Add Dialog Component`s story

* docs: Change storybook code docs

* refactor: Change unmount ref access

* docs: Add stopPropagation in story code
