+++ src/components/Dialog/Dialog.stories.tsx
@@ -1,17 +1,18 @@
 import type { Meta, StoryObj } from '@storybook/react';
 import { FC, useCallback, useMemo } from 'react';
 
+import { Button } from '@/components/Button/Button.example';
+import { clsx } from 'clsx';
 import { useDialog } from '.';
-import { DialogExample as DialogExampleBase } from './Dialog.example';
-import type { Props as DialogExampleBaseProps } from './Dialog.example';
+import { Dialog } from './Dialog';
+import type { Props as DialogProps } from './Dialog';
 
-type Props = Pick<DialogExampleBaseProps, 'className'> & {
-  shouldFocusTrap: boolean;
-  initialFocus: boolean;
+type Props = Pick<
+  DialogProps,
+  'className' | 'shouldFocusTrap' | 'initialFocus'
+> & {
   actionShowDialog: (value: unknown) => void;
-  actionCloseDialogMain: (value: unknown) => void;
-  actionCloseDialogSub: (value: unknown) => void;
-  actionClickAbort: (value: unknown) => void;
+  actionCloseDialog: (value: unknown) => void;
 };
 
 const DialogExample: FC<Props> = ({
@@ -19,20 +20,13 @@ const DialogExample: FC<Props> = ({
   shouldFocusTrap,
   initialFocus,
   actionShowDialog,
-  actionCloseDialogMain,
-  actionCloseDialogSub,
-  actionClickAbort,
+  actionCloseDialog,
 }) => {
   const { ref, isOpen, showDialog, closeDialog } = useDialog();
 
-  const shouldFocusTrapCal = useMemo(() => {
-    if (shouldFocusTrap) return undefined;
-    return false;
-  }, [shouldFocusTrap]);
-
   const initialFocusCal = useMemo(() => {
     if (initialFocus) return undefined;
-    return false;
+    return initialFocus;
   }, [initialFocus]);
 
   const handleShowDialog = useCallback(() => {
@@ -42,39 +36,67 @@ const DialogExample: FC<Props> = ({
 
   const handleCloseDialogMain = useCallback(() => {
     closeDialog();
-    actionCloseDialogMain('closeDialog');
-  }, [actionCloseDialogMain, closeDialog]);
+    actionCloseDialog('closeDialog click main');
+  }, [actionCloseDialog, closeDialog]);
 
   const handleCloseDialogSub = useCallback(() => {
     closeDialog();
-    actionCloseDialogSub('closeDialog');
-  }, [actionCloseDialogSub, closeDialog]);
+    actionCloseDialog('closeDialog click sub');
+  }, [actionCloseDialog, closeDialog]);
 
   const handleCloseDialogAway = useCallback(() => {
     closeDialog();
-    actionClickAbort('closeDialog');
-  }, [actionClickAbort, closeDialog]);
+    actionCloseDialog('closeDialog click away');
+  }, [actionCloseDialog, closeDialog]);
 
   return (
-    <DialogExampleBase
-      className={className}
-      ref={ref}
-      isOpen={isOpen}
-      shouldFocusTrap={shouldFocusTrapCal}
-      initialFocus={initialFocusCal}
-      handleShowDialog={handleShowDialog}
-      handleCloseDialogMain={handleCloseDialogMain}
-      handleCloseDialogSub={handleCloseDialogSub}
-      handleClickAway={handleCloseDialogAway}
-    />
+    <div>
+      <Button onClick={handleShowDialog}>showDialog</Button>
+      <Dialog
+        className={clsx('backdrop:bg-gray-900 backdrop:opacity-80', className)}
+        ref={ref}
+        isOpen={isOpen}
+        shouldFocusTrap={shouldFocusTrap}
+        initialFocus={initialFocusCal}
+        onClickAway={handleCloseDialogAway}
+      >
+        <div
+          className={clsx(
+            'h-80 w-96 p-2',
+            'flex flex-col justify-between',
+            'divide-y divide-solid',
+          )}
+        >
+          <div className={clsx('flex-initial', 'h-8')}>header</div>
+          <div className={clsx('flex-auto', 'overflow-y-scroll')}>
+            <div className={clsx('h-[500px]')}>main</div>
+          </div>
+          <div
+            className={clsx(
+              'flex-initial',
+              'h-max pt-2',
+              'flex items-center justify-between',
+            )}
+          >
+            <Button onClick={handleCloseDialogMain}>
+              closeDialog main!!!!
+            </Button>
+            <Button onClick={handleCloseDialogSub}>closeDialog sub!!!!</Button>
+          </div>
+        </div>
+      </Dialog>
+    </div>
   );
 };
 
-const code = `const DialogExample: FC = () => {
+const code = `import { Dialog, useDialog } from '@kamo88/react-dialog-hooks';
+
+const DialogExample = () => {
+
     const { ref, isOpen, showDialog, closeDialog } = useDialog();
   
     return (
-        <>
+        <div>
             <button type="button" onClick={showDialog}>showDialog</button>
             <Dialog
                 className="backdrop:bg-gray-900 backdrop:opacity-80"
@@ -90,17 +112,18 @@ const code = `const DialogExample: FC = () => {
                     <div>main</div>
                     <div>
                         footer 
-                        <button type="button" onClick={closeDialog}>closeDialog</button>
+                        <button type="button" onClick={closeDialog}>closeDialog main</button>
+                        <button type="button" onClick={closeDialog}>closeDialog sub</button>
                     </div>
                 </div>
             </Dialog>
-        </>
+        </div>
     );
   };
 `;
 
 const meta = {
-  title: 'components/Dialog/useDialog',
+  title: 'components/Dialog',
   component: DialogExample,
   parameters: {
     // Optional parameter to center the component in the Canvas. More info: https://storybook.js.org/docs/react/configure/story-layout
@@ -148,22 +171,8 @@ const meta = {
       See 'show code' for actual usage in this page.
     `,
     },
-    actionCloseDialogMain: {
-      action: 'actionCloseDialogMain',
-      description: `**Not Dialog Component Props**<br>
-    This is props for use in storybook.<br>
-    See 'show code' for actual usage in this page.
-  `,
-    },
-    actionCloseDialogSub: {
-      action: 'actionCloseDialogSub',
-      description: `**Not Dialog Component Props**<br>
-    This is props for use in storybook.<br>
-    See 'show code' for actual usage in this page.
-  `,
-    },
-    actionClickAbort: {
-      action: 'actionClickAbort',
+    actionCloseDialog: {
+      action: 'actionCloseDialog',
       description: `**Not Dialog Component Props**<br>
     This is props for use in storybook.<br>
     See 'show code' for actual usage in this page.
@@ -176,9 +185,4 @@ export default meta;
 
 type Story = StoryObj<typeof meta>;
 
-export const Dialog: Story = {
-  args: {
-    shouldFocusTrap: true,
-    initialFocus: true,
-  },
-};
+export const Default: Story = {};
